split index page into launch page components
