+++ components/launch-page/launch-page-break.tsx
@@ -1,0 +1,16 @@
+import { FC } from 'react';
+import cn from '~/styles/components/launch-page/launch-page-break.styl'
+
+
+const LaunchPageBreak: FC = () => (
+  <div className={cn.break}>
+    <div>
+      <hr/>
+    </div>
+    <div>
+      <hr/>
+    </div>
+  </div>
+)
+
+export default LaunchPageBreak
