+++ components/launch-page/launch-page-footer.tsx
@@ -1,0 +1,12 @@
+import { FC } from 'react';
+import cn from '~/styles/components/launch-page/launch-page-footer.styl'
+
+
+const LaunchPageFooter: FC = () => (
+  <footer className={cn.footer}>
+    <p>A BSA Open Tech project by Black Socialists in America</p>
+    <p>Coming in 2021</p>
+  </footer>
+)
+
+export default LaunchPageFooter
