+++ styles/components/launch-page/launch-page-break.styl
@@ -1,0 +1,22 @@
+@import '../../_breakpoints'
+ 
+.break
+  display: flex
+  flex-direction: row
+  div
+    flex-grow: 1
+
+    hr
+      border-top: 2px solid black
+
+  +above('laptop')
+    width: 100%;
+    justify-content: space-between
+
+    div
+      flex-grow: 0
+      
+    div:first-child
+      width: 54.94%
+    div:last-child
+      width: 42.22%
