+++ styles/components/launch-page/launch-page-footer.styl
@@ -1,0 +1,29 @@
+@import '../../_breakpoints'
+@import url('http://fonts.googleapis.com/css?family=Space+Mono')      
+.footer
+    margin-top: 45px
+    font-size: 10px
+    line-height: 9.5px
+    font-family: Space Mono, Helvetica
+
+    +above('laptop')
+      display: flex
+      justify-content: space-between
+      align-items: center
+      width: 100%
+      margin: 0 0 13px
+
+
+    p
+      +above('laptop')
+        line-height: 39px
+
+    p:first-child
+      margin-bottom: 23px
+
+      +above('laptop')
+        width: 56.32%
+        margin-bottom: 0
+        border-right: 2px solid black
+
+
