+++ styles/components/launch-page/launch-page-header.styl
@@ -1,0 +1,32 @@
+@import '../../_breakpoints'
+@import url('http://fonts.googleapis.com/css?family=Space+Mono')
+@import url('http://fonts.googleapis.com/css?family=Roboto+Mono')
+
+.header
+  display: flex
+  margin-bottom: 42px
+
+  figure
+    margin-right: 40px
+    height: 70px
+    width: 70px
+    background: url('/dual-power-logo--medium.png')
+
+    +above('laptop')
+      margin-right: 406px
+      margin-bottom: 14px
+      height: 58px
+      width: 58px
+      background: url('/dual-power-logo--small.png')
+
+  a
+    text-decoration: none;
+
+.social
+  flex-grow: 1
+  display: flex
+  flex-direction: column
+  justify-content: center
+  border-bottom: 2px solid #000000;
+  line-height: 22px
+  font-weight: 700
