+++ styles/components/launch-page/launch-page-power-section.styl
@@ -1,0 +1,43 @@
+@import '../../_breakpoints'
+@import url('http://fonts.googleapis.com/css?family=Space+Mono')
+@import url('http://fonts.googleapis.com/css?family=Roboto+Mono')
+
+.power
+  padding-bottom: 45px
+  border-top: 2px solid black
+
+  & > p
+    margin-bottom: 46px
+    width: 208px
+    padding: 13px
+    font-family: Space Mono, Helvetica
+    color: white
+    background-color: black
+
+    +above('laptop')
+      margin-bottom: 30px
+      width: 329px
+      height: 68px
+      padding: 25.24px 19px
+      font-family: Roboto Mono, Helvetica
+
+  & > div
+    line-height: 18.56px
+
+    +above('laptop')
+      display: flex
+
+      p
+        width: 341px
+
+  & > div:nth-child(2)
+    & > div:first-child
+      p
+        margin-bottom: 40px
+
+        +above('laptop')
+          margin-right: 49px
+    
+    & > div:nth-child(2)
+      p:first-child
+        margin-bottom: 16px
