+++ styles/components/launch-page/launch-page-title-section.styl
@@ -1,0 +1,39 @@
+@import '../../_breakpoints'
+@import url('http://fonts.googleapis.com/css?family=Space+Mono')
+  
+.title
+  margin-bottom: 44px
+
+  +above('laptop')
+    margin-right: 52px
+    width: 411px;
+    min-width: 411px;
+    height: auto
+
+  p:first-child
+    margin-bottom: 20px
+    font-size: 60px
+    line-height: 49px
+
+    +above('laptop')
+      margin-bottom: 19px
+      font-size: 80px
+      line-height: 64.8px
+
+  p:nth-child(2)
+    font-family: Space Mono, Helvetica
+    line-height: 28px
+
+.description
+  margin-bottom: 54px
+  line-height: 21.5px
+
+  +above('laptop')
+    font-size: 32px
+    line-height: 30px
+
+  p:first-child
+    margin-bottom: 8px
+
+    +above('laptop')
+      margin-bottom: 24px
