+++ styles/pages/index.styl
@@ -1,8 +1,4 @@
 @import '../_breakpoints'
-@import '../_theme'
-@import '../inputs'
-@import url('http://fonts.googleapis.com/css?family=Space+Mono')
-@import url('http://fonts.googleapis.com/css?family=Roboto+Mono')
 
 .index
   display: flex
@@ -22,181 +18,3 @@
     section:nth-child(2)
       +above('laptop')
         display: flex;
-
-
-.donate
-  position: absolute
-  right: 22px
-  top: 8px
-  display: flex
-  justify-content: center
-  align-items: center
-  border-radius: 50%
-  height: 83px;
-  width: 83px;
-  border: 2px solid black
-
-  +above('laptop')
-    right: 70px
-    top: 628px
-
-  &:hover
-    cursor: pointer;
-  
-
-.header
-  display: flex
-  margin-bottom: 42px
-
-  figure
-    margin-right: 40px
-    height: 70px
-    width: 70px
-    background: url('/dual-power-logo--medium.png')
-
-    +above('laptop')
-      margin-right: 406px
-      margin-bottom: 14px
-      height: 58px
-      width: 58px
-      background: url('/dual-power-logo--small.png')
-
-  a
-    text-decoration: none;
-
-.social
-  flex-grow: 1
-  display: flex
-  flex-direction: column
-  justify-content: center
-  border-bottom: 2px solid #000000;
-  line-height: 22px
-  font-weight: 700
-  
-.title
-  margin-bottom: 44px
-
-  +above('laptop')
-    margin-right: 52px
-    width: 411px;
-    min-width: 411px;
-    height: auto
-
-  p:first-child
-    margin-bottom: 20px
-    font-size: 60px
-    line-height: 49px
-
-    +above('laptop')
-      margin-bottom: 19px
-      font-size: 80px
-      line-height: 64.8px
-
-  p:nth-child(2)
-    font-family: Space Mono, Helvetica
-    line-height: 28px
-
-.description
-  margin-bottom: 54px
-  line-height: 21.5px
-
-  +above('laptop')
-    font-size: 32px
-    line-height: 30px
-
-  p:first-child
-    margin-bottom: 8px
-
-    +above('laptop')
-      margin-bottom: 24px
-
-.power
-  padding-bottom: 45px
-  border-top: 2px solid black
-
-  & > p
-    margin-bottom: 46px
-    width: 208px
-    padding: 13px
-    font-family: Space Mono, Helvetica
-    color: white
-    background-color: black
-
-    +above('laptop')
-      margin-bottom: 30px
-      width: 329px
-      height: 68px
-      padding: 25.24px 19px
-      font-family: Roboto Mono, Helvetica
-
-  & > div
-    line-height: 18.56px
-
-    +above('laptop')
-      display: flex
-
-      p
-        width: 341px
-
-  & > div:nth-child(2)
-    & > div:first-child
-      p
-        margin-bottom: 40px
-
-        +above('laptop')
-          margin-right: 49px
-    
-    & > div:nth-child(2)
-      p:first-child
-        margin-bottom: 16px
-
- 
-.break
-  display: flex
-  flex-direction: row
-  div
-    flex-grow: 1
-
-    hr
-      border-top: 2px solid black
-
-  +above('laptop')
-    width: 100%;
-    justify-content: space-between
-
-    div
-      flex-grow: 0
-      
-    div:first-child
-      width: 54.94%
-    div:last-child
-      width: 42.22%
- 
-      
-.footer
-    margin-top: 45px
-    font-size: 10px
-    line-height: 9.5px
-    font-family: Space Mono, Helvetica
-
-    +above('laptop')
-      display: flex
-      justify-content: space-between
-      align-items: center
-      width: 100%
-      margin: 0 0 13px
-
-
-    p
-      +above('laptop')
-        line-height: 39px
-
-    p:first-child
-      margin-bottom: 23px
-
-      +above('laptop')
-        width: 56.32%
-        margin-bottom: 0
-        border-right: 2px solid black
-
-
